[BE][S09P12A604-394] Feat: 아이템 조회 기능

+++ Back/wcc/api/src/main/java/com/ssafy/wcc/domain/member/db/entity/MemberItemPK.java
@@ -7,7 +7,7 @@ import java.io.Serializable;
 @Embeddable
 public class MemberItemPK implements Serializable {
     @Column(name="collection_id")
-    private Long collectionId;
+    private int collectionId;
 
     @Column(name="member_id")
     private Long memberId;
